requirements.txt
Flask==0.10.1
Flask-PyMongo==0.3.1
Flask-Cors==2.1.2
awesome-slugify==1.6.5
Flask-BasicAuth==0.2.0
unidecode
